# app/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AntCode Commando - E-Commerce Logistics API",
    "version": "2.0.0",
    "description": "Full CRUD over a normalized 5-table schema (customers, addresses, products, orders, payments) populated with Cameroonian volume data, plus a single idempotent /webhook/{provider} payment callback shared by MTN MoMo, Orange Money, Campay, Smobilpay, and GeniusPay."
  },
  "components": {
    "securitySchemes": {
      "providerSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Provider-Signature",
        "description": "Placeholder name for whichever signature header the {provider} path segment actually expects, resolved dynamically from _PROVIDER_CONFIG (app/routes/webhooks.py): momo -> X-Momo-Signature/MOMO_WEBHOOK_SECRET, orange -> X-Orange-Signature/ORANGE_WEBHOOK_SECRET, campay -> X-Campay-Signature/CAMPAY_WEBHOOK_SECRET, smobilpay -> X-Smobilpay-Signature/SMOBILPAY_WEBHOOK_SECRET, geniuspay -> X-Webhook-Signature/GENIUSPAY_WEBHOOK_SECRET. All of them are a hex-encoded HMAC-SHA256 of the raw request body except geniuspay, which signs f\"{X-Webhook-Timestamp}.{raw_request_body}\" instead (also requires the X-Webhook-Timestamp and X-Webhook-Event headers)."
      }
    },
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "order_id": { "type": "string", "example": "ECM-00001" },
          "customer_id": { "type": "integer" },
          "address_id": { "type": "integer" },
          "product_id": { "type": "integer" },
          "quantity": { "type": "integer", "example": 2 },
          "unit_price_fcfa": { "type": "integer", "example": 75000 },
          "neighborhood": { "type": "string", "example": "Akwa" },
          "city": { "type": "string", "example": "Douala" },
          "delivery_status": {
            "type": "string",
            "example": "Shipped",
            "enum": ["Pending", "Shipped", "Delivered"]
          },
          "payment_status": {
            "type": "string",
            "example": "Paid",
            "enum": ["Pending", "Paid", "Failed"]
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "product_id": { "type": "integer" },
          "name": { "type": "string", "example": "Smartphone Tecno Spark" },
          "category": { "type": "string", "example": "Electronics" },
          "unit_price_fcfa": { "type": "integer", "example": 95000 },
          "stock_quantity": { "type": "integer", "example": 50 }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "address_id": { "type": "integer" },
          "neighborhood": { "type": "string", "example": "Akwa" },
          "city": { "type": "string", "example": "Douala" },
          "street_details": { "type": "string", "nullable": true }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "customer_id": { "type": "integer" },
          "full_name": { "type": "string", "example": "Amina Njoya" },
          "phone_number": { "type": "string", "example": "+237690000001" },
          "addresses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Address" }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "per_page": { "type": "integer", "example": 20 },
          "total_records": { "type": "integer", "example": 2 },
          "total_pages": { "type": "integer", "example": 1 }
        }
      },
      "StandardProviderCallback": {
        "type": "object",
        "description": "Flat callback shape shared by momo, orange, campay, and smobilpay.",
        "required": ["order_id", "external_transaction_id", "amount_fcfa", "status"],
        "properties": {
          "order_id": { "type": "string", "example": "ECM-00001" },
          "external_transaction_id": { "type": "string", "example": "MOMO-TX-0001" },
          "amount_fcfa": { "type": "integer", "example": 15000 },
          "status": { "type": "string", "enum": ["SUCCESSFUL", "FAILED"] }
        }
      },
      "GeniusPayCallback": {
        "type": "object",
        "required": ["data"],
        "description": "The outcome (paid vs. failed) is carried by the X-Webhook-Event header (payment.success / payment.failed), not a body field.",
        "properties": {
          "data": {
            "type": "object",
            "required": ["transaction_id", "amount", "metadata"],
            "properties": {
              "transaction_id": { "type": "string", "example": "GENIUSPAY-TX-0001" },
              "amount": { "type": "integer", "example": 12000 },
              "metadata": {
                "type": "object",
                "required": ["order_id"],
                "properties": {
                  "order_id": { "type": "string", "example": "ECM-00001" }
                }
              }
            }
          }
        }
      },
      "WebhookCallback": {
        "description": "Body shape depends on which provider the {provider} path segment names: momo/orange/campay/smobilpay send the flat StandardProviderCallback shape, geniuspay sends the nested GeniusPayCallback shape.",
        "oneOf": [
          { "$ref": "#/components/schemas/StandardProviderCallback" },
          { "$ref": "#/components/schemas/GeniusPayCallback" }
        ]
      }
    }
  },
  "paths": {
    "/orders": {
      "get": {
        "summary": "List orders, optionally filtered by neighborhood and/or delivery status",
        "description": "Neighborhood filtering joins to addresses (idx_addresses_neighborhood_city + idx_orders_address_id); status filtering uses idx_orders_delivery_status. See docs/indexing_and_query_optimization_report.md.",
        "parameters": [
          {
            "name": "neighborhood",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "example": "Akwa"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["Pending", "Shipped", "Delivered"] },
            "example": "Shipped"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "1-indexed page number. Defaults to 1; clamped to 1 if lower or non-numeric.",
            "schema": { "type": "integer", "default": 1 },
            "example": 1
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "description": "Rows per page. Defaults to 20; clamped to 1 if lower or non-numeric.",
            "schema": { "type": "integer", "default": 20 },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "Matching orders, paginated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Order" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a single order",
        "description": "Accepts customer_id or customer_name+customer_phone+neighborhood(+city) (get-or-create), and product_id or product_name+category+unit_price_fcfa (get-or-create), plus quantity. order_id is auto-generated (ECM-NNNNN) if omitted.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "400": { "description": "Missing required field or invalid value (e.g. non-positive quantity)" }
        }
      }
    },
    "/orders/{id_or_ref}": {
      "get": {
        "summary": "Fetch a single order by order_id or transaction reference",
        "description": "app/routes/orders.py::order_detail: a string starting with \"ECM-\" resolves against the order's own order_id; anything else is looked up as a payment's external_transaction_id (e.g. a GeniusPay/MoMo/Orange transaction reference) via app/services/orders.py::get_order_by_transaction_reference, returning the order that payment was recorded against.",
        "parameters": [
          {
            "name": "id_or_ref",
            "in": "path",
            "required": true,
            "description": "order_id (e.g. \"ECM-00001\") or a payment transaction reference (e.g. \"MTX-A1B2C3D4E5\")",
            "schema": { "type": "string", "example": "ECM-00001" }
          }
        ],
        "responses": {
          "200": {
            "description": "The order",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "404": { "description": "No order matches that id_or_ref under either lookup rule" }
        }
      }
    },
    "/orders/{order_id}": {
      "put": {
        "summary": "Update an order's tracking state",
        "description": "Updates delivery_status and/or payment_status only -- customer/product/quantity are immutable after creation.",
        "parameters": [
          { "name": "order_id", "in": "path", "required": true, "schema": { "type": "string", "example": "ECM-00001" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "delivery_status": { "type": "string", "enum": ["Pending", "Shipped", "Delivered"] },
                  "payment_status": { "type": "string", "enum": ["Pending", "Paid", "Failed"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated order", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
          "400": { "description": "Invalid delivery_status or payment_status value" },
          "404": { "description": "order_id does not exist" }
        }
      },
      "delete": {
        "summary": "Delete an order",
        "parameters": [
          { "name": "order_id", "in": "path", "required": true, "schema": { "type": "string", "example": "ECM-00001" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "order_id does not exist" },
          "409": { "description": "order still has payments recorded against it" }
        }
      }
    },
    "/orders/sync": {
      "post": {
        "summary": "Bulk-ingest offline-captured orders",
        "description": "Idempotent on order_id: replaying an identical batch after a connectivity blackout skips every already-synced order instead of duplicating it.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } }
        },
        "responses": {
          "200": {
            "description": "Sync result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "synced": { "type": "integer" }, "skipped": { "type": "integer" } }
                }
              }
            }
          },
          "400": { "description": "Payload is not a JSON array, or an order is missing a required field" }
        }
      }
    },
    "/orders/{order_id}/checkout": {
      "post": {
        "summary": "Initiate a GeniusPay checkout session for an order",
        "parameters": [
          { "name": "order_id", "in": "path", "required": true, "schema": { "type": "string", "example": "ECM-00001" } }
        ],
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "checkout_url": { "type": "string" }, "transaction_reference": { "type": "string" } }
                }
              }
            }
          },
          "404": { "description": "order_id does not exist" },
          "502": { "description": "The outbound GeniusPay request failed" }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "List products (paginated)",
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "required": false, "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Matching products, paginated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a product",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
        },
        "responses": {
          "201": { "description": "Product created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "400": { "description": "Missing/invalid name, category, unit_price_fcfa, or stock_quantity" }
        }
      }
    },
    "/products/{product_id}": {
      "put": {
        "summary": "Partially update a product",
        "parameters": [
          { "name": "product_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
        },
        "responses": {
          "200": { "description": "Updated product", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "400": { "description": "Invalid unit_price_fcfa or stock_quantity" },
          "404": { "description": "product_id does not exist" }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "parameters": [
          { "name": "product_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "product_id does not exist" },
          "409": { "description": "product is referenced by an existing order" }
        }
      }
    },
    "/customers": {
      "post": {
        "summary": "Create a customer together with its first address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["full_name", "phone_number", "neighborhood", "city"],
                "properties": {
                  "full_name": { "type": "string" },
                  "phone_number": { "type": "string", "example": "+2376XXXXXXXX" },
                  "neighborhood": { "type": "string" },
                  "city": { "type": "string" },
                  "street_details": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Customer created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
          "400": { "description": "Missing required field or phone_number not in +2376XXXXXXXX format" },
          "409": { "description": "phone_number already registered to another customer" }
        }
      }
    },
    "/customers/{customer_id}": {
      "get": {
        "summary": "Fetch a customer with its addresses",
        "parameters": [
          { "name": "customer_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "The customer", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
          "404": { "description": "customer_id does not exist" }
        }
      },
      "put": {
        "summary": "Update a customer and/or its primary address",
        "parameters": [
          { "name": "customer_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Updated customer", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
          "400": { "description": "Invalid phone_number format" },
          "404": { "description": "customer_id does not exist" },
          "409": { "description": "phone_number already registered to another customer" }
        }
      },
      "delete": {
        "summary": "Delete a customer and its addresses",
        "parameters": [
          { "name": "customer_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "customer_id does not exist" },
          "409": { "description": "customer has existing orders" }
        }
      }
    },
    "/webhook/{provider}": {
      "post": {
        "summary": "Payment callback for the named provider",
        "description": "Single dynamic entry point for every provider's payment callback -- momo, orange, campay, smobilpay, and geniuspay. _webhook() (app/routes/webhooks.py) resolves that provider's own secret, signature header, and (geniuspay only) timestamp/event headers from _PROVIDER_CONFIG. Idempotent, keyed on (provider, external_transaction_id): replaying the same transaction id (e.g. after an MTN retry following a 3G timeout in Douala/Yaounde) returns already_processed instead of double-charging the order. For geniuspay, order_id is read from data.metadata.order_id, the order is marked Paid only when X-Webhook-Event is payment.success, and a webhook.test event is acknowledged without touching orders. Adding a new provider only means adding an entry to _PROVIDER_CONFIG -- not a new route.",
        "security": [{ "providerSignature": [] }],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "Which payment provider sent this callback. Any value with no entry in _PROVIDER_CONFIG returns 500.",
            "schema": {
              "type": "string",
              "enum": ["momo", "orange", "campay", "smobilpay", "geniuspay"]
            }
          },
          {
            "name": "X-Webhook-Timestamp",
            "in": "header",
            "required": false,
            "description": "geniuspay only: Unix timestamp GeniusPay signed alongside the body; part of the HMAC message.",
            "schema": { "type": "string", "example": "1700000000" }
          },
          {
            "name": "X-Webhook-Event",
            "in": "header",
            "required": false,
            "description": "geniuspay only: event type. payment.success marks the order Paid; webhook.test is acknowledged without touching orders; anything else leaves it unpaid.",
            "schema": { "type": "string", "example": "payment.success" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookCallback" }
            }
          }
        },
        "responses": {
          "200": { "description": "Processed, already processed (idempotent replay), or a geniuspay webhook.test acknowledgement" },
          "401": { "description": "Missing, invalid, or payload/timestamp-mismatched signature for the named provider" },
          "404": { "description": "order_id does not exist" },
          "500": { "description": "provider names a value with no entry in _PROVIDER_CONFIG (misconfiguration)" }
        }
      }
    }
  }
}